refactor: move widget details inputs in own components
