+++ src/modules/RA/DataLayer/components/tabs/WidgetsTab/WidgetDetails/WidgetGraphPreview.js
